benchmark/requirements.txt
# Bird Classification Benchmark System Requirements

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
PyYAML>=5.4.0
tqdm>=4.62.0
hydra-core>=1.2.0
omegaconf>=2.2.0

# Audio processing
librosa>=0.8.0
soundfile>=0.10.0

# Machine learning and metrics
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0

# BirdNET integration
birdnetlib>=0.8.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Optional: For improved performance
# numba>=0.56.0  # Speeds up librosa
# resampy>=0.2.2  # Better audio resampling 